backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis==5.2.0

# Celery
celery==5.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP clients
httpx==0.28.0
aiohttp==3.11.7

# AI Providers
google-generativeai==0.8.3
google-cloud-aiplatform==1.73.0
openai==1.57.2
anthropic==0.39.0

# Telegram
python-telegram-bot==21.7

# GitHub
PyGithub==2.5.0

# Document processing
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.0

# Linting
ruff==0.8.4
